# api/openapi/swagger.yaml
basePath: /
definitions:
...
  github_com_bsonger_devflow-service_internal_approute_domain.Route:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      environment_id:
        type: string
      host:
        type: string
      id:
        type: string
      name:
        type: string
      path:
        type: string
      service_name:
        type: string
      service_port:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_bsonger_devflow-service_internal_approute_domain.RouteInput:
    properties:
      application_id:
        type: string
      environment_id:
        type: string
      host:
        type: string
      name:
        type: string
      path:
        type: string
      service_name:
        type: string
      service_port:
        type: integer
    type: object
  github_com_bsonger_devflow-service_internal_approute_domain.RouteValidationResult:
    properties:
      errors:
        items:
          type: string
        type: array
      valid:
        type: boolean
    type: object
  github_com_bsonger_devflow-service_internal_appservice_domain.Service:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.ServicePort'
        type: array
      updated_at:
        type: string
    type: object
  github_com_bsonger_devflow-service_internal_appservice_domain.ServiceInput:
    properties:
      application_id:
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.ServicePort'
        type: array
    type: object
  github_com_bsonger_devflow-service_internal_appservice_domain.ServicePort:
    properties:
      name:
        type: string
      protocol:
        type: string
      service_port:
        type: integer
      target_port:
        type: integer
    type: object
...
  ? github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_approute_domain_Route
  : properties:
      data:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_approute_domain.Route'
    type: object
  ? github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_approute_domain_RouteValidationResult
  : properties:
      data:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_approute_domain.RouteValidationResult'
    type: object
  ? github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_appservice_domain_Service
  : properties:
      data:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.Service'
    type: object
...
  ? github_com_bsonger_devflow-service_internal_platform_httpx.ListResponse-github_com_bsonger_devflow-service_internal_approute_domain_Route
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_approute_domain.Route'
        type: array
      pagination:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.Pagination'
    type: object
  ? github_com_bsonger_devflow-service_internal_platform_httpx.ListResponse-github_com_bsonger_devflow-service_internal_appservice_domain_Service
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.Service'
        type: array
      pagination:
        $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.Pagination'
    type: object
...
  github_com_bsonger_devflow-service_internal_platform_httpx.Pagination:
    properties:
      enabled:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact: {}
  description: DevFlow meta-service HTTP API.
  title: DevFlow Meta Service API
  version: "1.0"
paths:
...
  /api/v1/routes:
    get:
      parameters:
      - description: Application ID
        in: query
        name: application_id
        type: string
      - description: Route name
        in: query
        name: name
        type: string
      - description: Include deleted items
        in: query
        name: include_deleted
        type: boolean
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ListResponse-github_com_bsonger_devflow-service_internal_approute_domain_Route'
      summary: List application routes
      tags:
      - Route
    post:
      consumes:
      - application/json
      parameters:
      - description: Route data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_approute_domain.RouteInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_approute_domain_Route'
      summary: Create application route
      tags:
      - Route
  /api/v1/routes/{route_id}:
    delete:
      parameters:
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: string
      - description: Application ID
        in: query
        name: application_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete application route
      tags:
      - Route
    patch:
      consumes:
      - application/json
      parameters:
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: string
      - description: Route data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_approute_domain.RouteInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Update application route
      tags:
      - Route
  /api/v1/routes:validate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Route data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_approute_domain.RouteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_approute_domain_RouteValidationResult'
      summary: Validate application route
      tags:
      - Route
  /api/v1/services:
    get:
      parameters:
      - description: Application ID
        in: query
        name: application_id
        type: string
      - description: Service name
        in: query
        name: name
        type: string
      - description: Include deleted items
        in: query
        name: include_deleted
        type: boolean
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.ListResponse-github_com_bsonger_devflow-service_internal_appservice_domain_Service'
      summary: List application services
      tags:
      - Service
    post:
      consumes:
      - application/json
      parameters:
      - description: Service data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.ServiceInput'
      - description: Service data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.ServiceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_bsonger_devflow-service_internal_platform_httpx.DataResponse-github_com_bsonger_devflow-service_internal_appservice_domain_Service'
      summary: Create application service
      tags:
      - Service
  /api/v1/services/{service_id}:
    delete:
      parameters:
      - description: Service ID
        in: path
        name: service_id
        required: true
        type: string
      - description: Application ID
        in: query
        name: application_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete application service
      tags:
      - Service
    patch:
      consumes:
      - application/json
      parameters:
      - description: Service ID
        in: path
        name: service_id
        required: true
        type: string
      - description: Service data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_bsonger_devflow-service_internal_appservice_domain.ServiceInput'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Update application service
      tags:
      - Service
...
swagger: "2.0"